source/requirements.txt
bcrypt==4.3.0
blinker==1.8.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==45.0.7
flask==3.0.3
Flask-Bcrypt==1.0.1
flask-sqlalchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
packaging==25.0
pycparser==2.22
pymysql==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.4
six==1.17.0
sqlalchemy==2.0.43
typing-extensions==4.13.2
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
